Improved partial key function
Rather than always take the upper bits, which are often all zeroed out,
we take an XOR of chunks of size partial_t of the hash, which produces a
properly compressed hash key. This is a somewhat expensive function, so
we also try to minimize the number of times the partial key is computed
by bundling it with the hash value.

+++ tests/unit-tests/test_minimum_load_factor.cc
@@ -7,7 +7,7 @@
 
 class BadHashFunction {
 public:
-    int operator()(int) {
+    size_t operator()(int) {
         return 0;
     }
 };
